Make function to collect information from resource chain

+++ Injected/structs.gen.h
@@ -5,7 +5,7 @@
 
 struct IslandResource
 {
-    int64_t type_id;
+    int32_t type_id;
     int32_t amount;
     int32_t capacity;
 };
